[#116] Renamed Transformer plugin to Templating plugin
