--- model/model-transformer-adapter/pom.xml
@@ -1,42 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>net.zscript</groupId>
-        <artifactId>zscript-model</artifactId>
-        <version>0.0.1-SNAPSHOT</version>
-    </parent>
-
-    <artifactId>zscript-model-transformer-adapter</artifactId>
-    <name>Zscript Model Generator Transformer Adapter</name>
-
-    <properties>
-        <guava.version>32.1.2-jre</guava.version>
-    </properties>
-
-    <dependencies>
-        <dependency>
-            <groupId>net.zscript</groupId>
-            <artifactId>transformer-maven-plugin-adapter</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>net.zscript</groupId>
-            <artifactId>zscript-model-schema</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>com.google.guava</groupId>
-            <artifactId>guava</artifactId>
-            <version>${guava.version}</version>
-        </dependency>
-    </dependencies>
-
-    <build>
-        <plugins>
-
-        </plugins>
-    </build>
-</project>
