+++ model/pom.xml
@@ -17,7 +17,7 @@
         <module>language-definition</module>
         <module>schema</module>
         <module>standard-module-definitions</module>
-        <module>model-transformer-adapter</module>
+        <module>model-templating-adapter</module>
     </modules>
 
     <dependencies>
