+++ receivers/jvm/java-model-components/pom.xml
@@ -19,13 +19,13 @@
         <plugins>
             <plugin>
                 <groupId>net.zscript</groupId>
-                <artifactId>transformer-maven-plugin</artifactId>
+                <artifactId>templating-maven-plugin</artifactId>
                 <version>${project.version}</version>
 
                 <dependencies>
                     <dependency>
                         <groupId>net.zscript</groupId>
-                        <artifactId>zscript-model-transformer-adapter</artifactId>
+                        <artifactId>zscript-model-templating-adapter</artifactId>
                         <version>${project.version}</version>
                     </dependency>
                     <dependency>
@@ -37,7 +37,7 @@
 
                 <configuration>
                     <outputDirectory>${project.build.directory}/generated-sources/zscript</outputDirectory>
-                    <transformMapperClass>net.zscript.model.transformer.adapter.ZscriptModuleTransformerPluginMapper</transformMapperClass>
+                    <templatingMapperClass>net.zscript.model.templating.adapter.ZscriptModuleTemplatingPluginMapper</templatingMapperClass>
                 </configuration>
 
                 <executions>
